fix: add posibility to create, update and delete properties for owners

+++ src/backend/routes/property_routes.py
@@ -1,11 +1,10 @@
-from backend.models import User
-from fastapi import APIRouter, Depends
+from backend.models import User, Role
+from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy.ext.asyncio import AsyncSession
-
-from src.backend.services import PropertyService
-from src.backend.schemas.properties_schemas import PropertiesUpdate, PropertiesCreate, PropertiesResponse
+from src.backend.core.dependencies import get_owner_or_admin_user
 from src.backend.database.database import get_session
-from src.backend.core.dependencies import get_admin_user
+from src.backend.schemas.properties_schemas import PropertiesUpdate, PropertiesCreate, PropertiesResponse
+from src.backend.services import PropertyService
 from typing import List
 
 properties_router = APIRouter(
@@ -16,13 +15,29 @@ properties_router = APIRouter(
 PropertyService = PropertyService()
 
 
+async def check_owner(
+        session: AsyncSession,
+        owner: User,
+        property_id: int
+) -> bool:
+    existing_property = await PropertyService.get_property(session=session, property_id=property_id)
+    if existing_property.owner_id == owner.user_id:
+        return True
+    raise HTTPException(
+        status_code=403,
+        detail="Access Denied",
+    )
+
+
 @properties_router.get("/{property_id}", response_model=PropertiesResponse)
 async def get_property(property_id: int, session: AsyncSession = Depends(get_session)):
+    """Отримати готель по id"""
     return await PropertyService.get_property(session=session, property_id=property_id)
 
 
 @properties_router.get("/", response_model=List[PropertiesResponse])
 async def get_all_properties(session: AsyncSession = Depends(get_session)):
+    """Отримати список всіх готелів"""
     return await PropertyService.get_all_properties(session=session)
 
 
@@ -30,9 +45,15 @@ async def get_all_properties(session: AsyncSession = Depends(get_session)):
 async def create_property(
         property_create: PropertiesCreate,
         session: AsyncSession = Depends(get_session),
-        admin: User = Depends(get_admin_user)
+        current_user: User = Depends(get_owner_or_admin_user)
 ):
-    return await PropertyService.create_property(session=session, properties_create=property_create)
+    """Створити власність, може тільки адмін або власник"""
+    if current_user.role in [Role.ADMIN, Role.OWNER]:
+        return await PropertyService.create_property(session=session, properties_create=property_create)
+    raise HTTPException(
+        status_code=403,
+        detail="Access Denied",
+    )
 
 
 @properties_router.put("/{property_id}", response_model=PropertiesResponse)
@@ -40,19 +61,27 @@ async def update_property(
         property_id: int,
         property_update: PropertiesUpdate,
         session: AsyncSession = Depends(get_session),
-        admin: User = Depends(get_admin_user)
+        current_user: User = Depends(get_owner_or_admin_user)
 ):
-    return await PropertyService.update_property(
-        session=session,
-        properties_update=property_update,
-        property_id=property_id
+    """"Оновити дані про власність"""
+    if await check_owner(session=session, owner=current_user, property_id=property_id):
+        return PropertyService.update_property(session=session, property_id=property_id, properties_update=property_update)
+    raise HTTPException(
+        status_code=403,
+        detail="Access Denied",
     )
 
 
-@properties_router.delete("/{property_id}", response_model=PropertiesResponse)
+@properties_router.delete("/{property_id}", response_model=PropertiesResponse, status_code=204)
 async def delete_property(
         property_id: int,
         session: AsyncSession = Depends(get_session),
-        admin: User = Depends(get_admin_user)
+        current_user: User = Depends(get_owner_or_admin_user)
 ):
-    return await PropertyService.delete_property(session=session, property_id=property_id)
+    """Видалити власність"""
+    if await check_owner(session=session, owner=current_user, property_id=property_id):
+        await PropertyService.delete_property(session=session, property_id=property_id)
+    raise HTTPException(
+        status_code=403,
+        detail="Access Denied",
+    )
